Tweak font weights & styles
