+++ lib/screens/home/home.dart
@@ -37,12 +37,7 @@ class _HomeState extends State<Home> {
                 'Search',
               ];
 
-              return Text(
-                titles[_homeStore.selectedIndex],
-                style: const TextStyle(
-                  fontWeight: FontWeight.w800,
-                ),
-              );
+              return Text(titles[_homeStore.selectedIndex]);
             },
           ),
           actions: [
