+++ lib/screens/settings/sections/account_settings.dart
@@ -1,9 +1,11 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:frosty/constants/constants.dart';
 import 'package:frosty/core/auth/auth_store.dart';
 import 'package:frosty/screens/settings/stores/settings_store.dart';
-import 'package:frosty/screens/settings/widgets/blocked_users.dart';
+import 'package:frosty/widgets/block_button.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 class AccountSettings extends StatelessWidget {
@@ -178,7 +180,10 @@ class ProfileCard extends StatelessWidget {
               leading: CircleAvatar(
                 foregroundImage: CachedNetworkImageProvider(authStore.user.details!.profileImageUrl),
               ),
-              title: Text(authStore.user.details!.displayName),
+              title: Text(
+                authStore.user.details!.displayName,
+                style: const TextStyle(fontWeight: FontWeight.w600),
+              ),
               trailing: OutlinedButton.icon(
                 onPressed: () => _showLogoutDialog(context),
                 icon: const Icon(Icons.logout_outlined),
@@ -202,6 +207,54 @@ class ProfileCard extends StatelessWidget {
             ),
           );
         },
+      ),
+    );
+  }
+}
+
+class BlockedUsers extends StatelessWidget {
+  final AuthStore authStore;
+
+  const BlockedUsers({
+    Key? key,
+    required this.authStore,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Blocked Users'),
+      ),
+      body: RefreshIndicator(
+        onRefresh: () async {
+          HapticFeedback.lightImpact();
+          await authStore.user.refreshBlockedUsers(headers: authStore.headersTwitch);
+        },
+        child: Observer(
+          builder: (context) {
+            if (authStore.user.blockedUsers.isEmpty) {
+              return const Center(
+                child: Text('You don\'t have any blocked users.'),
+              );
+            }
+            return ListView(
+              children: authStore.user.blockedUsers.map(
+                (user) {
+                  final displayName = regexEnglish.hasMatch(user.displayName) ? user.displayName : user.displayName + ' (${user.userLogin})';
+                  return ListTile(
+                    title: Text(displayName),
+                    trailing: BlockButton(
+                      authStore: authStore,
+                      targetUser: displayName,
+                      targetUserId: user.userId,
+                    ),
+                  );
+                },
+              ).toList(),
+            );
+          },
+        ),
       ),
     );
   }
