+++ lib/screens/settings/sections/chat_settings.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:frosty/constants/constants.dart';
 import 'package:frosty/screens/settings/stores/settings_store.dart';
+import 'package:frosty/widgets/section_header.dart';
 
 class ChatSettings extends StatefulWidget {
   final SettingsStore settingsStore;
@@ -83,13 +84,7 @@ class _ChatSettingsState extends State<ChatSettings> {
             ),
           ),
           const SizedBox(height: 10.0),
-          const Padding(
-            padding: EdgeInsets.fromLTRB(10.0, 20.0, 10.0, 5.0),
-            child: Text(
-              'Message Appearance',
-              style: TextStyle(fontWeight: FontWeight.w600, fontSize: 18.0),
-            ),
-          ),
+          const SectionHeader('Message Appearance'),
           ExpansionTile(
             title: const Text(
               'Preview',
