+++ lib/widgets/section_header.dart
@@ -4,12 +4,14 @@ class SectionHeader extends StatelessWidget {
   final String text;
   final EdgeInsets padding;
   final double fontSize;
+  final FontWeight fontWeight;
 
   const SectionHeader(
     this.text, {
     Key? key,
     this.padding = const EdgeInsets.fromLTRB(10.0, 20.0, 10.0, 5.0),
     this.fontSize = 18.0,
+    this.fontWeight = FontWeight.w600,
   }) : super(key: key);
 
   @override
@@ -20,7 +22,7 @@ class SectionHeader extends StatelessWidget {
         text,
         style: TextStyle(
           fontSize: fontSize,
-          fontWeight: FontWeight.bold,
+          fontWeight: fontWeight,
         ),
       ),
     );
